Moved util files to utils dir
