+++ vs_utils/utils/tests/nnscore_pdb_test.py
@@ -11,11 +11,11 @@ import shutil
 import unittest
 import numpy as np
 
-from vs_utils.features.nnscore_pdb import PDB
-from vs_utils.features.nnscore_utils import Point
-from vs_utils.features.nnscore_utils import Atom
-from vs_utils.features.nnscore_utils import average_point
-from vs_utils.features.tests import __file__ as test_directory
+from vs_utils.utils.nnscore_pdb import PDB
+from vs_utils.utils.nnscore_utils import Point
+from vs_utils.utils.nnscore_utils import Atom
+from vs_utils.utils.nnscore_utils import average_point
+from vs_utils.utils.tests import __file__ as test_directory
 
 
 def data_dir():
